Move all modal components into new Modals directory. Update various import statements, remove unnecessary comments
