+++ src/components/ContactForm/ContactForm.js
@@ -1,7 +1,6 @@
 import React from "react";
 import Button from "../Button/Button";
 import SelectDropdown from "./SelectDropdown/SelectDropdown";
-
 import "./ContactForm.css";
 
 export default function ContactForm() {
