+++ src/components/ImageComp/ImageComp.css
@@ -35,17 +35,21 @@
 }
 @media screen and (max-width: 950px) {
   .large {
-    margin: 0 -80px 0 0;
+    margin: 0 -80px 0 10px;
     transform: scale(0.8);
   }
   .small {
     margin: 0 -10px 0 0;
     transform: scale(0.9);
   }
+  .last {
+    margin-right: 0;
+    margin-left: 15px;
+  }
 }
 @media screen and (max-width: 850px) {
   .large {
-    margin: 0 -130px 0 0;
+    margin: 0 -80px 0 0;
     transform: scale(0.7);
   }
 }
