+++ src/components/ImageComp/ImageComp.scss
@@ -50,7 +50,7 @@
 
 @media screen and (max-width: 950px) {
   .large {
-    margin: 0 -80px 0 0;
+    margin: 0 -80px 0 10px;
     transform: scale(0.8);
   }
 
@@ -58,11 +58,15 @@
     margin: 0 -10px 0 0;
     transform: scale(0.9);
   }
+  .last {
+    margin-right: 0;
+    margin-left: 15px;
+  }
 }
 
 @media screen and (max-width: 850px) {
   .large {
-    margin: 0 -130px 0 0;
+    margin: 0 -80px 0 0;
     transform: scale(0.7);
   }
 }
