+++ src/components/Modals/AddToCartConfirmationModal/AddToCartConfirmationModal.js
@@ -1,27 +1,19 @@
 import React from "react";
-import { useSelector } from "react-redux";
-import useFetch from "../../hooks/useFetch";
-import { useParams } from "react-router-dom";
-import { Link } from "react-router-dom";
+import { Link, useParams } from "react-router-dom";
 import "./AddToCartConfirmationModal.css";
-import Button from "../Button/Button";
+
+//component imports
+import Button from "../../Button/Button";
+
+// utility imports
+import useFetch from "../../../hooks/useFetch";
+import { useSelector } from "react-redux";
+import { getCartQuantityAndPrice } from "../../../utils/shopping";
 
 export default function WaitlistModal(props) {
   const id = useParams().id;
   const { product } = useFetch(`/products/${id}?populate=*`);
   const products = useSelector((state) => state.cart.products);
-  // console.log(product);
-
-  //^ export to module
-  const getCartQuantityAndPrice = () => {
-    let totalQuantity = 0;
-    let totalPrice = 0;
-
-    for (let i = 0; i < products.length; i++) {
-      totalQuantity += products[i].itemQuantity;
-    }
-    return [totalQuantity, totalPrice];
-  };
 
   return (
     <div
@@ -30,7 +22,7 @@ export default function WaitlistModal(props) {
     >
       <div className='confirmation-wrapper'>
         <div className='confirmation-title-wrapper'>
-          {/* <h1>{`Added to Cart (${getCartQuantityAndPrice()[0]})`}</h1> */}
+          {/* <h1>{`Added to Cart (${getCartQuantityAndPrice(products)[0]})`}</h1> */}
           <h1>Added to Cart</h1>
         </div>
         <div className='confirmation-body-wrapper'>
