+++ src/components/Modals/ContactModal/ContactModal.scss
@@ -1,4 +1,4 @@
-@import "../styleFunctions/functions";
+@import "../../styleFunctions/functions";
 
 //modal class to gray out the page
 .contact-modal-page-wrapper {
