+++ src/components/Modals/FactoryModal/FactoryModal.scss
@@ -1,4 +1,4 @@
-@import "../styleFunctions/functions";
+@import "../../styleFunctions/functions";
 
 //! each country modals width is overwritten in Factories.scss
 .factory-modal-wrapper {
