+++ src/components/Modals/HamburgerMenuModal/HamburgerMenuModal.scss
@@ -1,4 +1,4 @@
-@import "../styleFunctions/functions";
+@import "../../styleFunctions/functions";
 
 .burger-menu-outer-wrapper {
   //^ set parent relative
