+++ src/components/Modals/ShoppingCartModal/ShoppingCartModal.scss
@@ -1,4 +1,4 @@
-@import "../../components/styleFunctions/functions";
+@import "../../styleFunctions/functions";
 
 //^This class exists for the purpose graying out the background when the modal is visible. Think of it as an invisible overlay of the entire page. This work in conjunction with the grey-out class at the bottom of this file
 .cart-modal-page-wrapper {
