+++ src/utils/shopping.js
@@ -15,6 +15,7 @@ export const updateProductQuantity = (operation, currentQuantity) => {
     return updatedQuantity;
   } else return;
 };
+
 //
 //
